% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_seurat}
\alias{aggregate_seurat}
\title{Aggregate expression values by groups}
\usage{
aggregate_seurat(object, by, slot = "counts", method = c("sum", "mean"))
}
\arguments{
\item{object}{A Seurat object}

\item{by}{A vector providing the names of the columns to be used as groups
stored in the \code{@metadata} slot. Last group listed corresponds to the
population of interest}

\item{slot}{Expression matrix to split by groups:
("counts", "data", or "scale.data")}

\item{method}{Sum or average expression values}
}
\value{
A list containing Seurat objects corresponding to each of the groups
}
\description{
Creates pseudo-bulk gene expression values based on sample groups
}
\examples{
data <- SeuratObject::pbmc_small
aggregate_seurat(data, by = c("groups", "RNA_snn_res.1"))
}
\author{
Jose Alquicira-Hernandez
}
