% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Braque-package.R
\docType{package}
\name{Braque-package}
\alias{Braque}
\alias{Braque-package}
\title{Braque: Analysis of meta samples from single cell data}
\description{
This package provides a set of methods to analyze meta samples from single cell data
}
\keyword{internal}
