% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{$,dge-pair-method}
\alias{$,dge-pair-method}
\title{Extractor for sample-level columns}
\usage{
\S4method{$}{`dge-pair`}(x, name)
}
\arguments{
\item{x}{\code{dge-pair} object}

\item{name}{Column name}
}
\value{
A vector containing column of interest
}
\description{
Extractor for sample-level columns for \code{dge-pair} object
}
\author{
Jose Alquicira Hernandez
}
