% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{colnames,dge-pair-method}
\alias{colnames,dge-pair-method}
\title{Retrieve sample names}
\usage{
\S4method{colnames}{`dge-pair`}(x)
}
\arguments{
\item{x}{\code{dge-pair} object}
}
\value{
A vector containing sample names
}
\description{
Retrieve sample names from \code{dge-pair} object
}
\author{
Jose Alquicira Hernandez
}
