% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{create_dge}
\alias{create_dge}
\title{\code{dge-pair} object constructor}
\usage{
create_dge(expr, meta, sample_col, condition_col)
}
\arguments{
\item{expr}{Data.frame with rows as genes and columns as samples/individuals
representing gene expression data}

\item{sample_col}{Column name in sample-level metadata containing sample ids}

\item{condition_col}{Column name in sample-level metadata containing the constrast,
for example: cell type population}

\item{metadata}{A data.table containing sample-level metadata}
}
\value{
\code{dge-pair} object
}
\description{
\code{dge-pair} object constructor
}
\author{
Jose Alquicira Hernandez
}
