% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{dge-pair}
\alias{dge-pair}
\alias{dge_pair}
\alias{dge-pair-class}
\title{Definition of 'dge-pair' class}
\description{
An S4 class to containing expression data and sample-level metadata
}
\section{Slots}{

\describe{
\item{\code{expr}}{Data.frame with rows as genes and columns as samples/individuals
representing gene expression data}

\item{\code{metadata}}{A data.table containing sample-level metadata}

\item{\code{sample_col}}{Column name in sample-level metadata containing sample ids}

\item{\code{condition_col}}{Column name in sample-level metadata containing the constrast,
for example: cell type population}
}}

