% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{expr,dge-pair-method}
\alias{expr,dge-pair-method}
\title{Retrieve expression data}
\usage{
\S4method{expr}{`dge-pair`}(object)
}
\arguments{
\item{object}{\code{dge-pair} object}
}
\value{
A data frame containing expression data
}
\description{
Retrieves expression data from dge-pair object
}
\author{
Jose Alquicira Hernandez
}
