% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_map_reduce.R
\name{map_matrix}
\alias{map_matrix}
\title{Map function to each group matrix}
\usage{
map_matrix(object, f)
}
\arguments{
\item{object}{A data.table object created with  \code{split_matrix}}

\item{f}{Function to apply to each expression matrix. Function must be
applied to genes (row-wise manner, e.g. Matrix::rowMeans()).}
}
\value{
A data.table containing an extra column called \code{result} containing
the output of the provided function
}
\description{
Applies function to each group matrix after running
\code{split_matrix}.
}
\examples{
data <- SeuratObject::pbmc_small
group_mats <- split_matrix(data, by = c("groups", "RNA_snn_res.1"))
group_mats <- map_matrix(group_mats, Matrix::rowMeans)
}
\author{
Jose Alquicira-Hernandez
}
