% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{metadata,dge-pair-method}
\alias{metadata,dge-pair-method}
\title{Retrieve metadata}
\usage{
\S4method{metadata}{`dge-pair`}(object)
}
\arguments{
\item{object}{\code{dge-pair} object}
}
\value{
A data.table containing sample-level metadata
}
\description{
Retrieves sample-level metadata from dge-pair object
}
\author{
Jose Alquicira Hernandez
}
