% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{names,dge-pair-method}
\alias{names,dge-pair-method}
\title{Retrieve sample-level column names}
\usage{
\S4method{names}{`dge-pair`}(x)
}
\arguments{
\item{x}{\code{dge-pair} object}
}
\value{
A vector containing ssample-level column names
}
\description{
Retrieve sample-level column namesfrom \code{dge-pair} object
}
\author{
Jose Alquicira Hernandez
}
