% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{prop}
\alias{prop}
\title{Retrieve proportion of gene expression per sample}
\usage{
prop(object)
}
\arguments{
\item{object}{\code{Seurat} object}
}
\value{
A data.frame containing proportion of gene expression per sample
}
\description{
Retrieve proportion of gene expression per sample from a
Seurat object created with \code{Braque}
}
\author{
Jose Alquicira Hernandez
}
