% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_map_reduce.R
\name{reduce_seurat}
\alias{reduce_seurat}
\title{Gather results across groups}
\usage{
reduce_seurat(object)
}
\arguments{
\item{object}{A data.table object created and processed with \code{split_matrix}
and \code{map_matrix}}
}
\value{
A list containing Seurat objects corresponding to each of the groups
}
\description{
Combines output results after running \code{map_matrix}. This
function assumes the results consist on output vectors generated from
applying a function operating in matrices. The grouping hierarchy in the
results follows the order provided via \code{by} when running
\code{split_matrix}.
}
\examples{
data <- SeuratObject::pbmc_small
group_mats <- split_matrix(data, by = c("groups", "RNA_snn_res.1"))
group_mats <- map_matrix(group_mats, Matrix::rowMeans)
reduce_seurat(group_mats)
}
\author{
Jose Alquicira-Hernandez
}
