% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{rownames,dge-pair-method}
\alias{rownames,dge-pair-method}
\title{Retrieve gene names}
\usage{
\S4method{rownames}{`dge-pair`}(x)
}
\arguments{
\item{x}{\code{dge-pair} object}
}
\value{
A vector containing gene names
}
\description{
Retrieve gene names from \code{dge-pair} object
}
\author{
Jose Alquicira Hernandez
}
