% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,dge-pair-method}
\alias{show,dge-pair-method}
\title{Print \code{dge-pair} object}
\usage{
\S4method{show}{`dge-pair`}(object)
}
\arguments{
\item{x}{\code{dge-pair} object}
}
\value{
Summary information about genes, samples, and contrast/condition
}
\description{
Print \code{dge-pair} object
}
\author{
Jose Alquicira Hernandez
}
