% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_map_reduce.R
\name{split_matrix}
\alias{split_matrix}
\title{Split expresion matrix by groups}
\usage{
split_matrix(object, by, slot = "counts")
}
\arguments{
\item{object}{A seurat object}

\item{by}{A vector providing the names of the columns to be used as groups
stored in the \code{@metadata} slot}

\item{slot}{Expression matrix to split by groups:
("counts", "data", or "scale.data")}
}
\value{
A tibble containing the group variables and the split matrices
}
\description{
Splits an expression matrix from a Seurat object based on
a set of columns stored in the metadata
}
\examples{
data <- Seurat::pbmc_small
split_matrix(data, by = c("groups", "RNA_snn_res.1"))
}
\author{
Jose Alquicira-Hernandez
}
