% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[,dge-pair-method}
\alias{[,dge-pair-method}
\title{Matrix-like accesor for genes and samples}
\usage{
\S4method{[}{`dge-pair`}(x, i, j, drop = FALSE)
}
\arguments{
\item{i}{row index (genes)}

\item{j}{column index (samples)}

\item{drop}{drop data.frame structure?}

\item{object}{\code{dge-pair} object}
}
\value{
A \code{dge-pair} object with subsetted data
}
\description{
Matrix-like accesor for genes and samples for dge-pair object
}
\author{
Jose Alquicira Hernandez
}
