% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[[,dge-pair-method}
\alias{[[,dge-pair-method}
\title{Extractor for sample-level columns}
\usage{
\S4method{[[}{`dge-pair`}(x, i)
}
\arguments{
\item{x}{\code{dge-pair} object}

\item{i}{Column names or indexes}
}
\value{
A data.table containing column of interest
}
\description{
Extractor for sample-level columns for \code{dge-pair} object
}
\author{
Jose Alquicira Hernandez
}
