% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{prop}
\alias{prop}
\title{Retrieve proportion of gene expression per sample}
\usage{
prop(object)
}
\description{
Retrieve proportion of gene expression per sample
}
\section{Functions}{
\itemize{
\item \code{prop}: Retrieve proportion of gene expression per sample
}}

